Update expected present time to optimize refresh rate
When the GPU/HWC composition cost too much time and complete in next VSYNC period,
the next frame will be skipped. But actually the SurfaceFlinger commit/present a
frame can be paralleled with previous frame composition(GPU3D or HWC). So we can
suppose that this frame may be present in 2 or more Vsync period.

Adding one more vsync period to expected present time for this frame if previous
frame does not present(present fence is not signaled). So this frame can prepare
early.
The expected present time that send to HWC still keep to use the original expected
present time to avoid HWC wait too loog time when present frame.

Change-Id: I3345234e3e9a257d75e4d6096be967711147e624
Signed-off-by: Bo Zhang <bo.zhang@nxp.com>